INSTALL.md: Add Virtualenv information
For those wanting to use Virtualenv, we need to make sure that the
`--system-site-packages` flag is passed in when creating it. Otherwise
the PyQt5 package won't be accessible since it needs to be installed
using Homebrew or Apt. Doing so installs PyQt5 to the system Python
where it isn't accessible from a normal Virtualenv.